[Part 1] Display logs for client and server (#128)
Clickup Ticket(s): https://app.clickup.com/t/8689qq767

I'm splitting this ticket into 2 PRs: one for back end and one for front end.

This is the back end one, where I am making 2 endpoints to retrieve the logs for the server and the clients of a job.

In order to make it possible, I had to add the location of the log paths in the filesystem to the database. The log file paths had to be returned from the functions that start both clients and servers, and also the client endpoint that starts a client. That approach may not work when we move this to slurm jobs, but we'll cross that bridge when we get there.