[Part 2] Display logs for client and server (#130)
Part 2: UI for displaying logs. Part 1: #128

Adding a link to the progress details section for both server and clients, which will open a modal that fetches and displays the logs (see screenshots below). The modal also has an option for refreshing the logs, which will trigger a refetch, and a download option, which will save the logs in a file and download it.