Adding FedProx as another model example (#161)
Clickup Ticket(s): https://app.clickup.com/t/868caa1b3

Adding MNIST with FedProx as another option for a model in order to have a more compelling demo.

Some refactorings were necessary to make it work because we were hardcoding the FedAvg strategy and the FlServer. Both of those are different for FedProx, so some things had to be moved around and made more flexible. Here is a summary of changes:

    Renamed florist.api.servers.common module to florist.api.servers.models
    Renamed MNIST model to MNIST with FedAvg
    Created a ServerFactory class that will provide a way to instantiate the "server constructor" for each one of the models.
    Moved the code that makes the strategy and the server to the get_fedavg_server, and created a get_fedprox_server function which will make FedProx's strategy and server
    Created FedProxConfigParser with the required config attributes for FedProx